21311LJ024 - Studierende, Abschlussprüfung, Land, Semester

The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (Education, Social Benefits, Health, Administration of Justice) -> 21 Bildung und Kultur (Education and Culture) -> 213 Hochschulen (Universities) -> 21311 Statistik der Studenten (Student Statistics).

df_21311LJ024

Format

A tibble with 79 rows and 4 variables:

BIL002

Studierende

BILAP3

Abschlussprüfung

DLAND

Land

SEMEST

Semester

Source

The data in this package was obtained from the Landesdatenbank NRW (retrieved on 2024-05-05) using the code '21311LJ024'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): Data licence Germany – attribution – Version 2.0. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: Compared to the raw data, the following changes were made: (1) technical columns ending in "QUAL", "ERR", or "LOCK" were removed, and (2) variable and value labels were added to each table using the labelled package. The code for all these transformations is available through this package.

Details

The following summary is automatically generated using skimr::skim()

Table: Data summary

Name	df_21311LJ024

Number of rows	79

Number of columns	4

_______________________

Column type frequency:

character	3

numeric	1

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

BILAP3	0	1	5	8	0	3	0

DLAND	0	1	2	2	0	1	0

SEMEST	0	1	10	10	0	31	0

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

BIL002	0	1	229634.2	196032	29	75092	149188	487673.5	529378	▇▅▁▁▆

Examples

library(labelled)

df_21311LJ024
#> # A tibble: 79 × 4
#>    BILAP3           DLAND                    SEMEST     BIL002
#>    <chr+lbl>        <chr+lbl>                <chr>       <dbl>
#>  1 HSAP1 [Bachelor] 05 [Nordrhein-Westfalen] WS 1999/00   1158
#>  2 HSAP1 [Bachelor] 05 [Nordrhein-Westfalen] WS 2000/01   3270
#>  3 HSAP1 [Bachelor] 05 [Nordrhein-Westfalen] WS 2001/02   7748
#>  4 HSAP1 [Bachelor] 05 [Nordrhein-Westfalen] WS 2002/03  16993
#>  5 HSAP1 [Bachelor] 05 [Nordrhein-Westfalen] WS 2003/04  29766
#>  6 HSAP1 [Bachelor] 05 [Nordrhein-Westfalen] WS 2004/05  38741
#>  7 HSAP1 [Bachelor] 05 [Nordrhein-Westfalen] WS 2005/06  59200
#>  8 HSAP1 [Bachelor] 05 [Nordrhein-Westfalen] WS 2006/07  87675
#>  9 HSAP1 [Bachelor] 05 [Nordrhein-Westfalen] WS 2007/08 138795
#> 10 HSAP1 [Bachelor] 05 [Nordrhein-Westfalen] WS 2008/09 194183
#> # ℹ 69 more rows