21311LJ185 - Studierende, Geschlecht, Land der Hochschulzugangsberechtigung, Land, Semester

The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (Education, Social Benefits, Health, Administration of Justice) -> 21 Bildung und Kultur (Education and Culture) -> 213 Hochschulen (Universities) -> 21311 Statistik der Studenten (Student Statistics).

df_21311LJ185

Format

A tibble with 1.115 rows and 5 variables:

BIL002

Studierende

GES

Geschlecht

BILZUBL

Land der Hochschulzugangsberechtigung

DLAND

Land

SEMEST

Semester

Source

The data in this package was obtained from the Landesdatenbank NRW (retrieved on 2024-05-05) using the code '21311LJ185'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): Data licence Germany – attribution – Version 2.0. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: Compared to the raw data, the following changes were made: (1) technical columns ending in "QUAL", "ERR", or "LOCK" were removed, and (2) variable and value labels were added to each table using the labelled package. The code for all these transformations is available through this package.

Details

The following summary is automatically generated using skimr::skim()

Table: Data summary

Name	df_21311LJ185

Number of rows	1115

Number of columns	5

_______________________

Column type frequency:

character	4

numeric	1

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

GES	0	1	4	4	0	2	0

BILZUBL	0	1	2	8	0	18	0

DLAND	0	1	2	2	0	1	0

SEMEST	0	1	10	10	0	31	0

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

BIL002	0	1	16270.04	49129.08	1	1097.5	2054	7786	287395	▇▁▁▁▁

Examples

library(labelled)

df_21311LJ185
#> # A tibble: 1,115 × 5
#>    GES             BILZUBL                 DLAND                   SEMEST BIL002
#>    <chr+lbl>       <chr+lbl>               <chr+lbl>               <chr>   <dbl>
#>  1 GESM [männlich] 01 [Schleswig-Holstein] 05 [Nordrhein-Westfale… WS 19…   1935
#>  2 GESM [männlich] 01 [Schleswig-Holstein] 05 [Nordrhein-Westfale… WS 19…   1988
#>  3 GESM [männlich] 01 [Schleswig-Holstein] 05 [Nordrhein-Westfale… WS 19…   2090
#>  4 GESM [männlich] 01 [Schleswig-Holstein] 05 [Nordrhein-Westfale… WS 19…   2144
#>  5 GESM [männlich] 01 [Schleswig-Holstein] 05 [Nordrhein-Westfale… WS 19…   2115
#>  6 GESM [männlich] 01 [Schleswig-Holstein] 05 [Nordrhein-Westfale… WS 19…   2154
#>  7 GESM [männlich] 01 [Schleswig-Holstein] 05 [Nordrhein-Westfale… WS 19…   2172
#>  8 GESM [männlich] 01 [Schleswig-Holstein] 05 [Nordrhein-Westfale… WS 19…   2139
#>  9 GESM [männlich] 01 [Schleswig-Holstein] 05 [Nordrhein-Westfale… WS 20…   2086
#> 10 GESM [männlich] 01 [Schleswig-Holstein] 05 [Nordrhein-Westfale… WS 20…   2067
#> # ℹ 1,105 more rows