21311LJ211 - Studierende, Fächergruppe, Altersgruppen (unter 18, 1-Jahres-Schritte, 60 und mehr), Land, Semester

The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (Education, Social Benefits, Health, Administration of Justice) -> 21 Bildung und Kultur (Education and Culture) -> 213 Hochschulen (Universities) -> 21311 Statistik der Studenten (Student Statistics).

Usage

df_21311LJ211

Format

A tibble with 9.692 rows and 5 variables:

BIL002

Studierende

BINWFG3

Fächergruppe

ALTNW08

Altersgruppen (unter 18, 1-Jahres-Schritte, 60 und mehr)

DLAND

Land

SEMEST

Semester

Source

The data in this package was obtained from the Landesdatenbank NRW (retrieved on 2024-05-05) using the code '21311LJ211'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): Data licence Germany – attribution – Version 2.0. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: Compared to the raw data, the following changes were made: (1) technical columns ending in "QUAL", "ERR", or "LOCK" were removed, and (2) variable and value labels were added to each table using the labelled package. The code for all these transformations is available through this package.

Details

The following summary is automatically generated using skimr::skim()

Table: Data summary

Name	df_21311LJ211

Number of rows	9692

Number of columns	5

_______________________

Column type frequency:

character	4

numeric	1

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

BINWFG3	0	1	4	4	0	8	0

ALTNW08	0	1	6	9	0	44	0

DLAND	0	1	2	2	0	1	0

SEMEST	0	1	10	10	0	28	0

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

BIL002	0	1	1718.37	3435.3	1	48	295	1454.25	25316	▇▁▁▁▁

Examples

library(labelled)

df_21311LJ211
#> # A tibble: 9,692 × 5
#>    BINWFG3                      ALTNW08                   DLAND    SEMEST BIL002
#>    <chr+lbl>                    <chr+lbl>                 <chr+lb> <chr>   <dbl>
#>  1 FG01 [Geisteswissenschaften] ALT000B18 [unter 18 Jahr… 05 [Nor… WS 19…      3
#>  2 FG01 [Geisteswissenschaften] ALT000B18 [unter 18 Jahr… 05 [Nor… WS 19…     43
#>  3 FG01 [Geisteswissenschaften] ALT000B18 [unter 18 Jahr… 05 [Nor… WS 19…      7
#>  4 FG01 [Geisteswissenschaften] ALT000B18 [unter 18 Jahr… 05 [Nor… WS 19…      5
#>  5 FG01 [Geisteswissenschaften] ALT000B18 [unter 18 Jahr… 05 [Nor… WS 19…      5
#>  6 FG01 [Geisteswissenschaften] ALT000B18 [unter 18 Jahr… 05 [Nor… WS 19…      2
#>  7 FG01 [Geisteswissenschaften] ALT000B18 [unter 18 Jahr… 05 [Nor… WS 20…      6
#>  8 FG01 [Geisteswissenschaften] ALT000B18 [unter 18 Jahr… 05 [Nor… WS 20…      4
#>  9 FG01 [Geisteswissenschaften] ALT000B18 [unter 18 Jahr… 05 [Nor… WS 20…     12
#> 10 FG01 [Geisteswissenschaften] ALT000B18 [unter 18 Jahr… 05 [Nor… WS 20…      9
#> # ℹ 9,682 more rows