21311LJ296 - Studienanfänger im 1. Hochschulsemester, Studienanfänger im 1. Fachsemester, Studienfach, Land, Studienjahr

The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (Education, Social Benefits, Health, Administration of Justice) -> 21 Bildung und Kultur (Education and Culture) -> 213 Hochschulen (Universities) -> 21311 Statistik der Studenten (Student Statistics).

df_21311LJ296

Format

A tibble with 5.762 rows and 5 variables:

BIL041

Studienanfänger im 1. Hochschulsemester

BIL042

Studienanfänger im 1. Fachsemester

BILNWSF1

Studienfach

DLAND

Land

SJAHR

Studienjahr

Source

The data in this package was obtained from the Landesdatenbank NRW (retrieved on 2024-05-05) using the code '21311LJ296'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): Data licence Germany – attribution – Version 2.0. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: Compared to the raw data, the following changes were made: (1) technical columns ending in "QUAL", "ERR", or "LOCK" were removed, and (2) variable and value labels were added to each table using the labelled package. The code for all these transformations is available through this package.

Details

The following summary is automatically generated using skimr::skim()

Table: Data summary

Name	df_21311LJ296

Number of rows	5762

Number of columns	5

_______________________

Column type frequency:

character	2

numeric	3

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

BILNWSF1	0	1	5	5	0	229	0

DLAND	0	1	2	2	0	1	0

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

SJAHR	0	1	2007.99	8.55	1993	2001.00	2008	2015.00	2022	▇▇▇▇▇

BIL041	0	1	477.29	1044.40	0	28.25	113	399.00	15100	▇▁▁▁▁

BIL042	0	1	776.21	1725.97	0	50.25	176	604.75	20287	▇▁▁▁▁

Examples

library(labelled)

df_21311LJ296
#> # A tibble: 5,762 × 5
#>    BILNWSF1            DLAND                    SJAHR BIL041 BIL042
#>    <chr+lbl>           <chr+lbl>                <dbl>  <dbl>  <dbl>
#>  1 SF001 [Ägyptologie] 05 [Nordrhein-Westfalen]  1993     33     62
#>  2 SF001 [Ägyptologie] 05 [Nordrhein-Westfalen]  1994     26     53
#>  3 SF001 [Ägyptologie] 05 [Nordrhein-Westfalen]  1995     32     64
#>  4 SF001 [Ägyptologie] 05 [Nordrhein-Westfalen]  1996     39     62
#>  5 SF001 [Ägyptologie] 05 [Nordrhein-Westfalen]  1997     46     99
#>  6 SF001 [Ägyptologie] 05 [Nordrhein-Westfalen]  1998     38     81
#>  7 SF001 [Ägyptologie] 05 [Nordrhein-Westfalen]  1999     37     75
#>  8 SF001 [Ägyptologie] 05 [Nordrhein-Westfalen]  2000     44     71
#>  9 SF001 [Ägyptologie] 05 [Nordrhein-Westfalen]  2001     38     85
#> 10 SF001 [Ägyptologie] 05 [Nordrhein-Westfalen]  2002     50    110
#> # ℹ 5,752 more rows