21311NW001B - Studierende / Studenten, Land, Hochschulen, Fachsemester, Nationalität, Hörerstatus, Angestrebte Abschlussprüfung, Semester

The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (Education, Social Benefits, Health, Administration of Justice) -> 21 Bildung und Kultur (Education and Culture) -> 213 Hochschulen (Universities) -> 21311 Statistik der Studenten (Student Statistics).

df_21311NW001B

Format

A tibble with 214.109 rows and 8 variables:

bil002

Studierende / Studenten

dland

Land

bilhs1

Hochschulen

bilfs1

Fachsemester

nat

Nationalität

bilhn1

Hörerstatus

bilap1

Angestrebte Abschlussprüfung

semest

Semester

Source

The data in this package was obtained from the Landesdatenbank NRW (retrieved on 2021-07-12) using the code '21311NW001B'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): Data licence Germany – attribution – Version 2.0. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: (1) column names were changed to lower case, (2) some technical columns were excluded and (3) value labels were turned into factors using the description of each value label. The code for all these transformations is available through this package.

Details

The following summary is automatically generated using skimr::skim() Table: Data summary

Name	df_21311NW001B

Number of rows	214109

Number of columns	8

_______________________

Column type frequency:

factor	7

numeric	1

________________________

Group variables	None

Variable type: factor

skim_variable	n_missing	complete_rate	ordered	n_unique	top_counts

dland	0	1	FALSE	1	Nor: 214109

bilhs1	0	1	FALSE	172	Uni: 13377, Uni: 12722, Uni: 11084, Tec: 9301

bilfs1	0	1	FALSE	20	1. : 18335, 3. : 16867, 5. : 15835, 7. : 14371

nat	0	1	FALSE	2	Deu: 133414, Aus: 80695

bilhn1	0	1	FALSE	2	Hau: 180817, Neb: 33292

bilap1	0	1	FALSE	59	Bac: 25225, Dip: 20206, Pro: 16443, Dip: 13514

semest	0	1	FALSE	22	WS : 11675, WS : 11581, WS : 11114, WS : 10778

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

bil002_val	0	1	60.91	213.05	1	2	7	34	11242	▇▁▁▁▁

Examples

df_21311NW001B

#> # A tibble: 214,109 x 8
#>    dland    bilhs1           bilfs1   nat   bilhn1 bilap1      semest bil002_val
#>    <fct>    <fct>            <fct>    <fct> <fct>  <fct>       <fct>       <dbl>
#>  1 Nordrhe… Hochschule Macr… 1. Fach… Ausl… Haupt… "Bachelor … WS 20…         14
#>  2 Nordrhe… Hochschule Macr… 1. Fach… Ausl… Haupt… "Bachelor … WS 20…         11
#>  3 Nordrhe… Hochschule Macr… 1. Fach… Ausl… Haupt… "Bachelor … WS 20…         20
#>  4 Nordrhe… Hochschule Macr… 1. Fach… Ausl… Haupt… "Master (F… WS 20…          1
#>  5 Nordrhe… Hochschule Macr… 1. Fach… Ausl… Haupt… "Abschluss… WS 20…          4
#>  6 Nordrhe… Hochschule Macr… 1. Fach… Ausl… Haupt… " angestre… WS 20…          3
#>  7 Nordrhe… Hochschule Macr… 1. Fach… Ausl… Haupt… " angestre… WS 20…          3
#>  8 Nordrhe… Hochschule Macr… 1. Fach… Deut… Haupt… "Bachelor … WS 20…        147
#>  9 Nordrhe… Hochschule Macr… 1. Fach… Deut… Haupt… "Bachelor … WS 20…        191
#> 10 Nordrhe… Hochschule Macr… 1. Fach… Deut… Haupt… "Bachelor … WS 20…        185
#> # … with 214,099 more rows