21311NW005B - Studierende / Studenten, Land, Hochschulen, Hochschulsemester, Nationalität, Hörerstatus, Angestrebte Abschlussprüfung, Semester

The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (Education, Social Benefits, Health, Administration of Justice) -> 21 Bildung und Kultur (Education and Culture) -> 213 Hochschulen (Universities) -> 21311 Statistik der Studenten (Student Statistics).

df_21311NW005B

Format

A tibble with 256.299 rows and 8 variables:

bil002

Studierende / Studenten

dland

Land

bilhs1

Hochschulen

bilsh1

Hochschulsemester

nat

Nationalität

bilhn1

Hörerstatus

bilap1

Angestrebte Abschlussprüfung

semest

Semester

Source

The data in this package was obtained from the Landesdatenbank NRW (retrieved on 2021-07-12) using the code '21311NW005B'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): Data licence Germany – attribution – Version 2.0. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: (1) column names were changed to lower case, (2) some technical columns were excluded and (3) value labels were turned into factors using the description of each value label. The code for all these transformations is available through this package.

Details

The following summary is automatically generated using skimr::skim() Table: Data summary

Name	df_21311NW005B

Number of rows	256299

Number of columns	8

_______________________

Column type frequency:

factor	7

numeric	1

________________________

Group variables	None

Variable type: factor

skim_variable	n_missing	complete_rate	ordered	n_unique	top_counts

dland	0	1	FALSE	1	Nor: 256299

bilhs1	0	1	FALSE	172	Uni: 15476, Uni: 14289, Uni: 11632, Uni: 10729

bilsh1	0	1	FALSE	20	20.: 17120, 11.: 15538, 9. : 15489, 7. : 15175

nat	0	1	FALSE	2	Deu: 159876, Aus: 96423

bilhn1	0	1	FALSE	2	Hau: 219318, Neb: 36981

bilap1	0	1	FALSE	59	Bac: 35247, Dip: 23048, Mas: 20507, Pro: 15125

semest	0	1	FALSE	22	WS : 13751, WS : 13748, WS : 13388, WS : 13141

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

bil002_val	0	1	50.88	186.56	1	2	6	26	15507	▇▁▁▁▁

Examples

df_21311NW005B

#> # A tibble: 256,299 x 8
#>    dland     bilhs1      bilsh1    nat    bilhn1  bilap1      semest  bil002_val
#>    <fct>     <fct>       <fct>     <fct>  <fct>   <fct>       <fct>        <dbl>
#>  1 Nordrhei… Alanus Hoc… 1. Hochs… Auslä… Haupth… Abschlussp… WS 200…          1
#>  2 Nordrhei… Alanus Hoc… 1. Hochs… Auslä… Haupth… Abschlussp… WS 201…          7
#>  3 Nordrhei… Alanus Hoc… 1. Hochs… Auslä… Haupth… Abschlussp… WS 201…          1
#>  4 Nordrhei… Alanus Hoc… 1. Hochs… Auslä… Haupth… Abschlussp… WS 201…          3
#>  5 Nordrhei… Alanus Hoc… 1. Hochs… Auslä… Haupth… Abschlussp… WS 201…          1
#>  6 Nordrhei… Alanus Hoc… 1. Hochs… Auslä… Haupth… Abschlussp… WS 201…          4
#>  7 Nordrhei… Alanus Hoc… 1. Hochs… Auslä… Haupth… Bachelor (… WS 200…          8
#>  8 Nordrhei… Alanus Hoc… 1. Hochs… Auslä… Haupth… Bachelor (… WS 201…          3
#>  9 Nordrhei… Alanus Hoc… 1. Hochs… Auslä… Haupth… Bachelor (… WS 201…         15
#> 10 Nordrhei… Alanus Hoc… 1. Hochs… Auslä… Haupth… Bachelor (… WS 201…         11
#> # … with 256,289 more rows