21311NW007A - Studierende / Studenten, Land, Hochschulen, Studienfach, Geschlecht, Semester

The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (Education, Social Benefits, Health, Administration of Justice) -> 21 Bildung und Kultur (Education and Culture) -> 213 Hochschulen (Universities) -> 21311 Statistik der Studenten (Student Statistics).

df_21311NW007A

Format

A tibble with 53.814 rows and 6 variables:

bil002

Studierende / Studenten

dland

Land

bilhs1

Hochschulen

bilsf1

Studienfach

ges

Geschlecht

semest

Semester

Source

The data in this package was obtained from the Landesdatenbank NRW (retrieved on 2021-07-12) using the code '21311NW007A'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): Data licence Germany – attribution – Version 2.0. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: (1) column names were changed to lower case, (2) some technical columns were excluded and (3) value labels were turned into factors using the description of each value label. The code for all these transformations is available through this package.

Details

The following summary is automatically generated using skimr::skim() Table: Data summary

Name	df_21311NW007A

Number of rows	53814

Number of columns	6

_______________________

Column type frequency:

factor	5

numeric	1

________________________

Group variables	None

Variable type: factor

skim_variable	n_missing	complete_rate	ordered	n_unique	top_counts

dland	0	1	FALSE	1	Nor: 53814

bilhs1	0	1	FALSE	172	Uni: 3341, Uni: 3244, Uni: 3012, Uni: 2720

bilsf1	0	1	FALSE	234	Wir: 1968, Bet: 1756, Mas: 1323, Ele: 1276

ges	0	1	FALSE	2	män: 26953, wei: 26861

semest	0	1	FALSE	22	WS : 3356, WS : 3260, WS : 3162, WS : 3054

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

bil002_val	0	1	242.35	519.73	1	24	89	266	16714	▇▁▁▁▁

Examples

df_21311NW007A

#> # A tibble: 53,814 x 6
#>    dland      bilhs1                    bilsf1          ges   semest  bil002_val
#>    <fct>      <fct>                     <fct>           <fct> <fct>        <dbl>
#>  1 Nordrhein… Hochschule Macromedia fü… Betriebswirtsc… männ… WS 201…          6
#>  2 Nordrhein… Hochschule Macromedia fü… Betriebswirtsc… weib… WS 201…         10
#>  3 Nordrhein… Hochschule Macromedia fü… Film und Ferns… männ… WS 201…         95
#>  4 Nordrhein… Hochschule Macromedia fü… Film und Ferns… männ… WS 201…         96
#>  5 Nordrhein… Hochschule Macromedia fü… Film und Ferns… männ… WS 201…        103
#>  6 Nordrhein… Hochschule Macromedia fü… Film und Ferns… weib… WS 201…         42
#>  7 Nordrhein… Hochschule Macromedia fü… Film und Ferns… weib… WS 201…         47
#>  8 Nordrhein… Hochschule Macromedia fü… Film und Ferns… weib… WS 201…         42
#>  9 Nordrhein… Hochschule Macromedia fü… Graphikdesign/… männ… WS 201…         19
#> 10 Nordrhein… Hochschule Macromedia fü… Graphikdesign/… männ… WS 201…         24
#> # … with 53,804 more rows