Metadata containing the variables for each student statistics table

The metadata for each table name in tables is sourced using wiesbaden::retrieve_metadata(). Data source is the Landesdatenbank NRW.

Usage

metadata

Format

A tibble with 1.881 rows and 4 variables:

tablename

Table name

name

Variable name

description

Variable description

unit

Unit for each measure

Source

The data in this package was obtained from the Landesdatenbank NRW (retrieved on 2024-05-05). The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): Data licence Germany – attribution – Version 2.0. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Examples

metadata
#> # A tibble: 1,881 × 4
#>    table      name   description              unit    
#>    <chr>      <chr>  <chr>                    <chr>   
#>  1 21311LJ001 BIL002 Studierende              "Anzahl"
#>  2 21311LJ001 BIL040 Ausländische Studierende "Anzahl"
#>  3 21311LJ001 BIL045 Nebenhörer               "Anzahl"
#>  4 21311LJ001 BIL046 Beurlaubte               "Anzahl"
#>  5 21311LJ001 DLAND  Land                     ""      
#>  6 21311LJ001 SEMEST Semester                 ""      
#>  7 21311LJ002 BIL002 Studierende              "Anzahl"
#>  8 21311LJ002 BIL040 Ausländische Studierende "Anzahl"
#>  9 21311LJ002 BIL045 Nebenhörer               "Anzahl"
#> 10 21311LJ002 BIL046 Beurlaubte               "Anzahl"
#> # ℹ 1,871 more rows