Student statistics tables ("21311 Statistik der Studenten")

The table list is sourced using wiesbaden::retrieve_datalist(). Data source is the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (Education, Social Benefits, Health, Administration of Justice) -> 21 Bildung und Kultur (Education and Culture) -> 213 Hochschulen (Universities) -> 21311 Statistik der Studenten (Student Statistics).

tables

Format

A tibble with 326 rows and 2 variables:

tablename

Table name

description

Description

Source

The data in this package was obtained from the Landesdatenbank NRW (retrieved on 2021-07-11) using the code 21311. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): Data licence Germany – attribution – Version 2.0. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Examples

tables
#> # A tibble: 326 × 2
#>    tablename  description                                                       
#>    <chr>      <chr>                                                             
#>  1 21311LJ001 Studierende, Ausländische Studierende, Nebenhörer, Beurlaubte, La…
#>  2 21311LJ002 Studierende, Ausländische Studierende, Nebenhörer, Beurlaubte, Ge…
#>  3 21311LJ003 Studierende, Nationalität, Land, Semester                         
#>  4 21311LJ004 Studierende, Geschlecht, Nationalität, Land, Semester             
#>  5 21311LJ005 Studierende, Nebenhörer, Hochschulart, Land, Semester             
#>  6 21311LJ006 Studierende, Nebenhörer, Geschlecht, Hochschulart, Land, Semester 
#>  7 21311LJ007 Studierende, Nationalität, Hochschulart, Land, Semester           
#>  8 21311LJ008 Studierende, Prüfungsgruppe, Land, Semester                       
#>  9 21311LJ009 Studierende, Prüfungsgruppe (nach Lehrämtern), Land, Semester     
#> 10 21311LJ010 Studierende, Hochschulsemester, Land, Semester                    
#> # ℹ 316 more rows