Value labels for all variables in the student statistics tables

The value labels are sourced using wiesbaden::retrieve_valuelabel(). Data source is the Landesdatenbank NRW.

Usage

value_labels

Format

A tibble with 1.056 rows and 3 variables:

variable

Variable

value_label

Value label (the name is misleading, this is effectively an ID)

description

The description for each value label (this is the text for each ID)

Source

The data in this package was obtained from the Landesdatenbank NRW (retrieved on 2024-05-05). The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): Data licence Germany – attribution – Version 2.0. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: (1) column names were changed to lower case.

Examples

value_labels
#> # A tibble: 1,056 × 3
#>    variable value_label description          
#>    <chr>    <chr>       <chr>                
#>  1 ALTNW08  ALT000B18   unter 18 Jahre       
#>  2 ALTNW08  ALT018      18 bis unter 19 Jahre
#>  3 ALTNW08  ALT019      19 bis unter 20 Jahre
#>  4 ALTNW08  ALT020      20 bis unter 21 Jahre
#>  5 ALTNW08  ALT021      21 bis unter 22 Jahre
#>  6 ALTNW08  ALT022      22 bis unter 23 Jahre
#>  7 ALTNW08  ALT023      23 bis unter 24 Jahre
#>  8 ALTNW08  ALT024      24 bis unter 25 Jahre
#>  9 ALTNW08  ALT025      25 bis unter 26 Jahre
#> 10 ALTNW08  ALT026      26 bis unter 27 Jahre
#> # ℹ 1,046 more rows