Metadata containing the variables for each student statistics table

The metadata for each table name in tables is sourced using wiesbaden::retrieve_metadata(). Data source is the Landesdatenbank NRW.

metadata

Format

A tibble with 154 rows and 4 variables:

tablename

Table name

name

Variable name

description

Variable description

unit

Unit for each measure

Source

The data in this package was obtained from the Landesdatenbank NRW (retrieved on 2021-07-11). The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): Data licence Germany – attribution – Version 2.0. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Examples

metadata

#> # A tibble: 154 x 4
#>    tablename   name   description                  unit    
#>    <chr>       <chr>  <chr>                        <chr>   
#>  1 21311LS001A BIL002 Studierende / Studenten      "Anzahl"
#>  2 21311LS001A DLAND  Land                         ""      
#>  3 21311LS001A BILHS1 Hochschulen                  ""      
#>  4 21311LS001A BILSF1 Studienfach                  ""      
#>  5 21311LS001A BILFS1 Fachsemester                 ""      
#>  6 21311LS001A BILAP1 Angestrebte Abschlussprüfung ""      
#>  7 21311LS001A SEMEST Semester                     ""      
#>  8 21311LS002A BIL002 Studierende / Studenten      "Anzahl"
#>  9 21311LS002A DLAND  Land                         ""      
#> 10 21311LS002A BILHS1 Hochschulen                  ""      
#> # … with 144 more rows