Additional information and comments for some variables

The additional information for each variable is sourced using wiesbaden::retrieve_varinfo(). Data source is the Landesdatenbank NRW.

variable_info

Format

A tibble with 16 rows and 2 variables:

variable

Variable name

description

Additional information for each variable

Source

The data in this package was obtained from the Landesdatenbank NRW (retrieved on 2024-05-05). The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): Data licence Germany – attribution – Version 2.0. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Examples

variable_info
#> # A tibble: 16 × 2
#>    variable description                                                         
#>    <chr>    <chr>                                                               
#>  1 BIL002   "Studierende sind in einem Fachstudium immatrikulierte\n(eingeschri…
#>  2 BIL041   "Studienanfänger/-innen im ersten Hochschulsemester sind\nStudieren…
#>  3 BIL042   "Studienanfänger/-innen im ersten Fachsemester sind\nStudierende, d…
#>  4 BIL045   "Studierende, die zugleich als Haupthörer/-innen an einer\nanderen …
#>  5 BILFS2   "Fachsemester sind die in einem Studiengang verbrachten\nSemester (…
#>  6 BILHS3   "Als Hochschulen werden alle staatlichen und nach Landesrecht\naner…
#>  7 BILHS5   "Seit dem Sommersemester 2017 wird in der Studierenden- und\nPrüfun…
#>  8 BILNWSF1 "Zentrales Erhebungsmerkmal der Studierenden- und\nPrüfungsstatisti…
#>  9 BILPG    "Für die Darstellung von aggregierten Ergebnissen der\nStudierenden…
#> 10 BILPG1   "Für die Darstellung von aggregierten Ergebnissen der\nStudierenden…
#> 11 BILSH2   "Die Angabe „Hochschulsemester“ gibt die Gesamtzahl der\nSemester a…
#> 12 BINWFG3  "Für die Darstellung von aggregierten Ergebnissen der\nStudierenden…
#> 13 BINWSB1  "Für die Darstellung von aggregierten Ergebnissen der\nStudierenden…
#> 14 DLAND    "Nordrhein-Westfalen"                                               
#> 15 HSLART   "Unter Hochschulart versteht man eine für Auswertungszwecke\ngebild…
#> 16 SJAHR    "Sommersemester und darauf folgendes Wintersemester"